refactor: move git from ext to client extension provider
